requirements.txt
alembic==1.4.0
aniso8601==6.0.0
appdirs==1.4.3
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.3.0
auth0-python==3.9.1
Authlib==0.14.1
awscli==1.18.110
bcrypt==3.1.7
beautifulsoup4==4.9.0
blinker==1.4
boto3==1.14.33
boto3-stubs==1.14.33.0
botocore==1.17.33
CacheControl==0.12.6
cachetools==4.1.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
cryptography==2.8
distlib==0.3.0
docutils==0.15.2
ecdsa==0.13.2
filelock==3.0.12
firebase-admin==4.3.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.7
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
flask-paginate==0.5.5
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-Threads==0.0.1a0
Flask-WTF==0.14.3
future==0.17.1
google-api-core==1.22.0
google-api-python-client==1.10.0
google-auth==1.20.0
google-auth-httplib2==0.0.4
google-cloud-core==1.4.1
google-cloud-firestore==1.8.1
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-resumable-media==0.7.0
googleapis-common-protos==1.52.0
grpcio==1.31.0
gunicorn==20.0.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.3.12
html5lib==1.0.1
httplib2==0.18.1
httpx==0.12.0
hyperframe==5.2.0
idna==2.9
importlib-metadata==1.5.0
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.10.0
joblib==0.16.0
lazy-object-proxy==1.4.0
livereload==2.6.3
lunr==0.5.8
lxml==4.5.0
Mako==1.1.1
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
mkdocs==1.1.2
mock==4.0.1
more-itertools==8.2.0
msgpack==1.0.0
mypy-boto3==1.14.33.0
nexmo==2.4.0
nltk==3.5
packaging==20.4
passlib==1.7.2
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.0
paypalrestsdk==1.13.1
Pillow==7.1.2
pipenv==2018.11.26
pluggy==0.13.1
protobuf==3.12.4
psycopg2-binary==2.8.2
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
pytest==5.3.5
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
python-ffmpeg-video-streaming==0.1.5
python-jose==3.1.0
pytz==2019.1
PyYAML==5.3.1
qrcode==6.1
regex==2020.7.14
requests==2.23.0
requests-cache-latest==0.4.12
rfc3986==1.3.2
rsa==4.0
s3transfer==0.3.3
six==1.12.0
sniffio==1.1.0
soupsieve==2.0
SQLAlchemy==1.3.4
tornado==6.0.4
tqdm==4.48.2
typed-ast==1.3.5
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==20.0.13
virtualenv-clone==0.5.3
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.15.2
wrapt==1.11.1
WTForms==2.2.1
zipp==3.1.0
